[#0] Put components on separated folders and add styled file for styled components
