+++ retro-board-ui/src/components/Column/Column.jsx
@@ -1,27 +1,7 @@
 import React from "react";
-import styled from "styled-components";
-import Task from "./task";
+import {Container, Title, TaskList} from './Column.styled'
+import Task from "../Task/Task";
 import { Droppable } from "react-beautiful-dnd";
-
-const Container = styled.div`
-  margin: 8px;
-  border: 1px solid lightgrey;
-  border-radius: 2px;
-  width: 220px;
-
-  display: flex;
-  flex-direction: column;
-`;
-const Title = styled.h3`
-  padding: 8px;
-`;
-const TaskList = styled.div`
-  padding: 8px;
-  transition: background-color 0.2 ease;
-  background-color: ${(props) => (props.isDraggingOver ? "skyblue" : "white")};
-  flex-grow: 1;
-  min-height: 100px;
-`;
 
 const Column = (props) => {
   return (
