+++ retro-board-ui/src/components/Task/Task.styled.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  border: 1px solid lightgrey;
+  border-radius: 2px;
+  padding: 8px;
+  margin-bottom: 8px;
+  background-color: ${(props) => (props.isDragging ? "lightgreen" : "white")};
+`;
+
+export const ReactionContainer = styled.div`
+  margin-right: 4px;
+`;
